requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
httpx==0.27.2
tldextract==5.1.2
beautifulsoup4==4.12.3
lxml==5.3.0
dnspython==2.7.0
python-whois==0.9.4
python-dotenv==1.0.1
pydantic==2.9.2
rapidfuzz==3.9.7